requirements.txt
flake8
ipython
matplotlib
numpy
PyYAML
pytest
scipy
sklearn
tensorflow-gpu
